move path matching and registration in separate file "path.go"
